Andro Goat APK
The andro goat apk is an open source vulnerable testing android application. 

The app is programmed with design and security flaws that allow researchers, students, hobbyists to exploit, learn, and test it in a safe environment. This is useful for training and learning possible security design mistakes in android apps and avoiding them.